lara/col: port Lara_Col_Hang to TRX
