+++ src/libtrx/include/libtrx/game/lara/col.h
@@ -4,7 +4,6 @@
 
 void Lara_Col_Update(ITEM *item, COLL_INFO *coll);
 
-extern void Lara_Col_Hang(ITEM *item, COLL_INFO *coll);
 extern void Lara_Col_ClimbLeft(ITEM *item, COLL_INFO *coll);
 extern void Lara_Col_ClimbRight(ITEM *item, COLL_INFO *coll);
 extern void Lara_Col_ClimbStance(ITEM *item, COLL_INFO *coll);
